+++ library/graphics_api/src/QueryPool.cpp
@@ -23,7 +23,7 @@ void QueryPool::get_result(std::span<float> out, const u32 first) const
 {
    std::vector<u64> timestamps{};
    timestamps.resize(out.size());
-   vkGetQueryPoolResults(m_queryPool.parent(), *m_queryPool, first, timestamps.size(), sizeof(u64) * timestamps.size(), timestamps.data(),
+   vkGetQueryPoolResults(m_queryPool.parent(), *m_queryPool, first, static_cast<u32>(timestamps.size()), sizeof(u64) * timestamps.size(), timestamps.data(),
                          sizeof(u64), VK_QUERY_RESULT_64_BIT | VK_QUERY_RESULT_WAIT_BIT);
 
    std::ranges::transform(timestamps, out.begin(),
@@ -34,7 +34,7 @@ float QueryPool::get_difference(const u32 begin, const u32 end) const
 {
    std::vector<u64> timestamps{};
    timestamps.resize(1 + end - begin);
-   vkGetQueryPoolResults(m_queryPool.parent(), *m_queryPool, begin, timestamps.size(), sizeof(u64) * timestamps.size(), timestamps.data(),
+   vkGetQueryPoolResults(m_queryPool.parent(), *m_queryPool, begin, static_cast<u32>(timestamps.size()), sizeof(u64) * timestamps.size(), timestamps.data(),
                          sizeof(u64), VK_QUERY_RESULT_64_BIT | VK_QUERY_RESULT_WAIT_BIT);
 
    return static_cast<float>(timestamps[timestamps.size() - 1] - timestamps[0]) * m_timestampPeriod / 1000000.0f;
